+++ test/unit/state/get-drag-impact/over-home-list/small-initial-movement-with-no-impact.spec.js
